+++ src/main/java/woowacrew/feed/controller/FeedApiController.java
@@ -1,18 +1,16 @@
 package woowacrew.feed.controller;
 
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import woowacrew.feed.dto.*;
+import woowacrew.feed.dto.FeedArticleResponseDto;
+import woowacrew.feed.dto.FeedSourceRequestDto;
+import woowacrew.feed.dto.FeedSourceResponseDto;
+import woowacrew.feed.dto.FeedSourceUpdateRequestDto;
 import woowacrew.feed.service.FeedService;
 
 import javax.validation.Valid;
 import java.util.List;
-
-import static woowacrew.article.free.service.ArticleInternalService.DEFAULT_ARTICLE_PAGE_SIZE;
 
 @RestController
 @RequestMapping("/api/feeds")
@@ -21,12 +19,6 @@ public class FeedApiController {
 
     public FeedApiController(FeedService feedService) {
         this.feedService = feedService;
-    }
-
-    @GetMapping
-    public ResponseEntity<FeedArticleResponseDtos> list
-            (@PageableDefault(size = DEFAULT_ARTICLE_PAGE_SIZE, sort = "publishedDate", direction = Sort.Direction.DESC) Pageable pageable) {
-        return ResponseEntity.ok(feedService.findAllFeedArticles(pageable));
     }
 
     @PostMapping
